Convert `progress.js` to TypeScript (#705)
